ECS: Add working collision rect with int MINOR

+++ src/ECS/ECSCustomTypes.hpp
@@ -22,8 +22,8 @@ extern "C"
 namespace Types {
 
     struct CollisionRect {
-            float width;
-            float height;
+            int width;
+            int height;
 
             NLOHMANN_DEFINE_TYPE_INTRUSIVE(CollisionRect, width, height);
     };
